+++ seed/fastapi/exhaustive/skip-formatting/resources/types/resources/union/types/animal.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 from .dog import Dog as resources_types_resources_union_types_dog_Dog
 from .cat import Cat as resources_types_resources_union_types_cat_Cat
 import typing
-import typing_extensions
 import datetime as dt
 from ......core.datetime_utils import serialize_datetime
 
@@ -30,7 +29,7 @@ class Animal(pydantic.BaseModel):
         if self.__root__.animal == "cat":
             return cat(
             resources_types_resources_union_types_cat_Cat(**self.__root__.dict(exclude_unset=True, exclude={"animal"})))
-    __root__: typing_extensions.Annotated[typing.Union[_Animal.Dog, _Animal.Cat], pydantic.Field(discriminator="animal")]
+    __root__: typing.Annotated[typing.Union[_Animal.Dog, _Animal.Cat], pydantic.Field(discriminator="animal")]
     def json(self, **kwargs: typing.Any) -> str:
         kwargs_with_defaults: typing.Any = { "by_alias": True, "exclude_unset": True, **kwargs }
         return super().json(**kwargs_with_defaults)
@@ -42,11 +41,11 @@ class Animal(pydantic.BaseModel):
         json_encoders = {dt.datetime: serialize_datetime}
 class _Animal:
     class Dog(resources_types_resources_union_types_dog_Dog):
-        animal: typing_extensions.Literal["dog"]
+        animal: typing.Literal["dog"]
         class Config:
             allow_population_by_field_name = True
     class Cat(resources_types_resources_union_types_cat_Cat):
-        animal: typing_extensions.Literal["cat"]
+        animal: typing.Literal["cat"]
         class Config:
             allow_population_by_field_name = True
 Animal.update_forward_refs()