+++ seed/fastapi/trace/resources/submission/types/test_case_grade.py
@@ -5,8 +5,6 @@ from __future__ import annotations
 import datetime as dt
 import typing
 
-import typing_extensions
-
 from ....core.datetime_utils import serialize_datetime
 from .test_case_hidden_grade import TestCaseHiddenGrade
 from .test_case_non_hidden_grade import TestCaseNonHiddenGrade
@@ -43,7 +41,7 @@ class TestCaseGrade(pydantic.BaseModel):
         if self.__root__.type == "nonHidden":
             return non_hidden(TestCaseNonHiddenGrade(**self.__root__.dict(exclude_unset=True, exclude={"type"})))
 
-    __root__: typing_extensions.Annotated[
+    __root__: typing.Annotated[
         typing.Union[_TestCaseGrade.Hidden, _TestCaseGrade.NonHidden], pydantic.Field(discriminator="type")
     ]
 
@@ -62,13 +60,13 @@ class TestCaseGrade(pydantic.BaseModel):
 
 class _TestCaseGrade:
     class Hidden(TestCaseHiddenGrade):
-        type: typing_extensions.Literal["hidden"]
+        type: typing.Literal["hidden"]
 
         class Config:
             allow_population_by_field_name = True
 
     class NonHidden(TestCaseNonHiddenGrade):
-        type: typing_extensions.Literal["nonHidden"]
+        type: typing.Literal["nonHidden"]
 
         class Config:
             allow_population_by_field_name = True