+++ seed/pydantic/trace/src/seed/trace/resources/commons/variable_value.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import typing
 
-import typing_extensions
-
 from .binary_tree_node_value import BinaryTreeNodeValue
 from .binary_tree_value import BinaryTreeValue
 from .doubly_linked_list_node_value import DoublyLinkedListNodeValue
@@ -21,65 +19,65 @@ except ImportError:
 
 
 class VariableValue_IntegerValue(pydantic.BaseModel):
-    type: typing_extensions.Literal["integerValue"]
+    type: typing.Literal["integerValue"]
     value: int
 
 
 class VariableValue_BooleanValue(pydantic.BaseModel):
-    type: typing_extensions.Literal["booleanValue"]
+    type: typing.Literal["booleanValue"]
     value: bool
 
 
 class VariableValue_DoubleValue(pydantic.BaseModel):
-    type: typing_extensions.Literal["doubleValue"]
+    type: typing.Literal["doubleValue"]
     value: float
 
 
 class VariableValue_StringValue(pydantic.BaseModel):
-    type: typing_extensions.Literal["stringValue"]
+    type: typing.Literal["stringValue"]
     value: str
 
 
 class VariableValue_CharValue(pydantic.BaseModel):
-    type: typing_extensions.Literal["charValue"]
+    type: typing.Literal["charValue"]
     value: str
 
 
 class VariableValue_MapValue(MapValue):
-    type: typing_extensions.Literal["mapValue"]
+    type: typing.Literal["mapValue"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class VariableValue_ListValue(pydantic.BaseModel):
-    type: typing_extensions.Literal["listValue"]
+    type: typing.Literal["listValue"]
     value: typing.List[VariableValue]
 
 
 class VariableValue_BinaryTreeValue(BinaryTreeValue):
-    type: typing_extensions.Literal["binaryTreeValue"]
+    type: typing.Literal["binaryTreeValue"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class VariableValue_SinglyLinkedListValue(SinglyLinkedListValue):
-    type: typing_extensions.Literal["singlyLinkedListValue"]
+    type: typing.Literal["singlyLinkedListValue"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class VariableValue_DoublyLinkedListValue(DoublyLinkedListValue):
-    type: typing_extensions.Literal["doublyLinkedListValue"]
+    type: typing.Literal["doublyLinkedListValue"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class VariableValue_NullValue(pydantic.BaseModel):
-    type: typing_extensions.Literal["nullValue"]
+    type: typing.Literal["nullValue"]
 
 
 VariableValue = typing.Union[