+++ seed/pydantic/trace/src/seed/trace/resources/problem/problem_description_board.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import typing
 
-import typing_extensions
-
 try:
     import pydantic.v1 as pydantic  # type: ignore
 except ImportError:
@@ -13,17 +11,17 @@ except ImportError:
 
 
 class ProblemDescriptionBoard_Html(pydantic.BaseModel):
-    type: typing_extensions.Literal["html"]
+    type: typing.Literal["html"]
     value: str
 
 
 class ProblemDescriptionBoard_Variable(pydantic.BaseModel):
-    type: typing_extensions.Literal["variable"]
+    type: typing.Literal["variable"]
     value: VariableValue
 
 
 class ProblemDescriptionBoard_TestCaseId(pydantic.BaseModel):
-    type: typing_extensions.Literal["testCaseId"]
+    type: typing.Literal["testCaseId"]
     value: str