+++ seed/pydantic/trace/src/seed/trace/resources/submission/exception_v_2.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import typing
 
-import typing_extensions
-
 from .exception_info import ExceptionInfo
 
 try:
@@ -15,14 +13,14 @@ except ImportError:
 
 
 class ExceptionV2_Generic(ExceptionInfo):
-    type: typing_extensions.Literal["generic"]
+    type: typing.Literal["generic"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class ExceptionV2_Timeout(pydantic.BaseModel):
-    type: typing_extensions.Literal["timeout"]
+    type: typing.Literal["timeout"]
 
 
 ExceptionV2 = typing.Union[ExceptionV2_Generic, ExceptionV2_Timeout]