+++ seed/pydantic/trace/src/seed/trace/resources/submission/submission_request.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import typing
 
-import typing_extensions
-
 from .initialize_problem_request import InitializeProblemRequest
 from .stop_request import StopRequest
 from .submit_request_v_2 import SubmitRequestV2
@@ -18,32 +16,32 @@ except ImportError:
 
 
 class SubmissionRequest_InitializeProblemRequest(InitializeProblemRequest):
-    type: typing_extensions.Literal["initializeProblemRequest"]
+    type: typing.Literal["initializeProblemRequest"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class SubmissionRequest_InitializeWorkspaceRequest(pydantic.BaseModel):
-    type: typing_extensions.Literal["initializeWorkspaceRequest"]
+    type: typing.Literal["initializeWorkspaceRequest"]
 
 
 class SubmissionRequest_SubmitV2(SubmitRequestV2):
-    type: typing_extensions.Literal["submitV2"]
+    type: typing.Literal["submitV2"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class SubmissionRequest_WorkspaceSubmit(WorkspaceSubmitRequest):
-    type: typing_extensions.Literal["workspaceSubmit"]
+    type: typing.Literal["workspaceSubmit"]
 
     class Config:
         allow_population_by_field_name = True
 
 
 class SubmissionRequest_Stop(StopRequest):
-    type: typing_extensions.Literal["stop"]
+    type: typing.Literal["stop"]
 
     class Config:
         allow_population_by_field_name = True